+++ Core/Input/InputDevice.h
@@ -1,11 +1,11 @@
 #pragma once
-
+#include <stdint.h>
 class InputDevice
 {
 public:
 	InputDevice() {}
 	~InputDevice() {}
 
-	virtual int CheckForRawInputSupport() = 0;
+	virtual int64_t CheckForRawInputSupport() = 0;
 };
